% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BinMat-class.R
\docType{methods}
\name{BinMat-class}
\alias{BinMat}
\alias{BinMat-class}
\alias{print}
\title{An S4 class to represent a matrix model for naive bayes spatial
data mining model.}
\usage{
BinMat(name_ID, DMNB, BMNB, ...)

\S4method{print}{BinMat}(x, ...)
}
\description{
An S4 class to represent a matrix model for naive bayes spatial
data mining model.
}
\section{Slots}{

\describe{
\item{\code{name_ID}}{a data.frame. rownames(name_ID) = colnames(DMNB) = colnames(BMNB).
A data containing the ID's for column names in slots DMNB and BMNB.}

\item{\code{DMNB}}{a matrix of integers. Contains the number of observations per cells
per column. The number of columns must be identical to the number of rows
in name_ID.}

\item{\code{BMNB}}{a logical matrix. Contains TRUE for all DMNB > 0, FALSE otherwise.
The number of columns must be identical to the number of rows in name_ID.}
}}
\examples{

# Good example
name_ID<-data.frame(name = c("carlos", "pepe"), row.names = c("X1","X2"), stringsAsFactors = FALSE)
DMNB<-matrix(rbinom(10,10, 0.5), ncol = nrow(name_ID), dimnames = list(1:5,rownames(name_ID)))
BMNB<-matrix(rbinom(10,1, 0.5), ncol = nrow(name_ID), dimnames = list(1:5,rownames(name_ID)))
x.mat<-BinMat(name_ID = name_ID, DMNB = DMNB, BMNB = BMNB)

x.mat
print(x.mat)

# Bad example
\dontrun{
name_ID<-data.frame(name = c("carlos", "pepe"), row.names = c("X1","X2"),
                    stringsAsFactors = F)
DMNB<-matrix(rep(1,10), ncol = nrow(name_ID), dimnames = list(1:5,rownames(name_ID)))
BMNB<-matrix(rep(1,10), ncol = 5, dimnames = list(1:2,c(LETTERS[1:5])))
x.mat<-BinMat(name_ID = name_ID, DMNB = DMNB, BMNB = BMNB)}

}
\author{
Enrique Del Callejo Canal (\email{edelcallejoc@gmail.com}).
}

