% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BinMatCount-class.R
\docType{methods}
\name{BinMatCount-class}
\alias{.BinMatCount}
\alias{BinMatCount}
\alias{BinMatCount-Contstructor}
\alias{BinMatCount-class}
\alias{new-BinMatCount}
\alias{print-BinMatCount}
\alias{show-BinMatCount}
\title{An S4 subclass from BinMat class.}
\usage{
BinMatCount(name_ID, DMNB, BMNB, Count, ...)

\S4method{show}{BinMatCount}(object)

\S4method{print}{BinMatCount}(x, ...)
}
\description{
An S4 subclass from BinMat class.
}
\details{
The elements of the list are defined as follow.

    \strong{N} an integer. It corresponds to the number of rows in the slot
    name_ID of BinMat object.

    \strong{Nc} a numeric vector.

    \strong{Nx} a numeric vector.

    \strong{Ncx} a numeric matrix.
}
\section{Slots}{

\describe{
\item{\code{Count}}{a list object whit 4 elements, N, Nc, Nx and Ncx. See details}
}}
\examples{

name_ID<-data.frame(name = c("carlos", "pepe"), row.names = c("X1","X2"), stringsAsFactors = FALSE)
DMNB<-matrix(rbinom(10,10, 0.5), ncol = nrow(name_ID), dimnames = list(1:5,rownames(name_ID)))
BMNB<-matrix(rbinom(10,1, 0.5), ncol = nrow(name_ID), dimnames = list(1:5,rownames(name_ID)))
x.mat<-BinMat(name_ID = name_ID, DMNB = DMNB, BMNB = BMNB)

x.mat

Count.mat<-list(N=10L, Nc = t(t(c(1L, 2L))), Nx = t(t(c(1L, 2L, 3L))), Ncx = matrix(1L:6L,2,3))

c.mat<-BinMatCount(name_ID = name_ID, DMNB = DMNB, BMNB = BMNB, Count = Count.mat)
}
\author{
Enrique Del Callejo Canal (\email{edelcallejoc@gmail.com}).
}

