% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BinMatEps-class.R
\docType{methods}
\name{BinMatEps-class}
\alias{BinMatEps}
\alias{BinMatEps-class}
\alias{Contstructor}
\alias{print}
\title{An S4 subclass from BinMat class.}
\usage{
BinMatEps(name_ID, DMNB, BMNB, Epsilon, ...)

\S4method{print}{BinMatEps}(x, ...)
}
\description{
An S4 subclass from BinMat class.
}
\details{
The elements of the list are defined as follow.

    \strong{Ecx}: numeric matrix. With s rows and m columns

    \strong{Pcnx}a numeric matrix. With s rows and m columns.
}
\section{Slots}{

\describe{
\item{\code{Epsilon}}{a list object whit 2 elements, Ecx and Encx.}
}}
\examples{

name_ID<-data.frame(name = c("carlos", "pepe"), row.names = c("X1","X2"), stringsAsFactors = F)
DMNB<-matrix(rbinom(10,10, 0.5), ncol = nrow(name_ID), dimnames = list(1:5,rownames(name_ID)))
BMNB<-matrix(rbinom(10,1, 0.5), ncol = nrow(name_ID), dimnames = list(1:5,rownames(name_ID)))

BinMatProb(name_ID = name_ID, DMNB = DMNB, BMNB = BMNB, Prob = list(Pc = matrix(1:2,2),
           Px = matrix(1:3,3), Pcx = matrix(1:6,2,3), Pcnx = matrix(1:6,2,3),
           Pxc = matrix(1:6,3,2), Pxnc = matrix(1:6,3,2), Pnxc = matrix(1:6,3,2),
            Pnxnc = matrix(1:6,3,2)))

}
\author{
Enrique Del Callejo Canal (\email{edelcallejoc@gmail.com}).
}

