% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BinMatProb-class.R
\docType{methods}
\name{BinMatProb-class}
\alias{BinMatProb}
\alias{BinMatProb-class}
\alias{Contstructor}
\alias{print}
\title{An S4 subclass from BinMat class.}
\usage{
BinMatProb(name_ID, DMNB, BMNB, Prob, ...)

\S4method{print}{BinMatProb}(x, ...)
}
\description{
An S4 subclass from BinMat class.
}
\details{
The elements of the list are defined as follow.

    \strong{Pc} a numeric matrix. With s rows and 1 column.

    \strong{Px} a numeric matrix. With m rows and 1 column.

    \strong{Pcx} a numeric matrix. With s rows and m columns.

    \strong{Pcnx}a numeric matrix. With s rows and m columns.

    \strong{Pxc} a numeric matrix. With m rows and s columns.

    \strong{Pxnc} a numeric matrix. With m rows and s columns.

    \strong{Pnxc} a numeric matrix. With m rows and s columns.

    \strong{Pnxnc} a numeric matrix. With m rows and s columns.
}
\section{Slots}{

\describe{
\item{\code{Prob}}{a list object whit 8 elements, Pc, Px, Pcx, Pcnx, Pxc, Pxnc,
Pnxc and Pnxnc. See details.}
}}
\examples{

name_ID<-data.frame(name = c("carlos", "pepe"), row.names = c("X1","X2"), stringsAsFactors = F)
DMNB<-matrix(rbinom(10,10, 0.5), ncol = nrow(name_ID), dimnames = list(1:5,rownames(name_ID)))
BMNB<-matrix(rbinom(10,1, 0.5), ncol = nrow(name_ID), dimnames = list(1:5,rownames(name_ID)))

BinMatProb(name_ID = name_ID, DMNB = DMNB, BMNB = BMNB, Prob = list(Pc = matrix(1:2,2),
           Px = matrix(1:3,3), Pcx = matrix(1:6,2,3), Pcnx = matrix(1:6,2,3),
           Pxc = matrix(1:6,3,2), Pxnc = matrix(1:6,3,2), Pnxc = matrix(1:6,3,2),
            Pnxnc = matrix(1:6,3,2)))

}
\author{
Enrique Del Callejo Canal (\email{edelcallejoc@gmail.com}).
}

