% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BinMatScore-class.R
\docType{methods}
\name{BinMatScore-class}
\alias{BinMatScore}
\alias{BinMatScore-class}
\alias{Contstructor}
\alias{print}
\title{An S4 subclass from BinMat class.}
\usage{
BinMatScore(name_ID, DMNB, BMNB, Score, ...)

\S4method{print}{BinMatScore}(x, ...)
}
\description{
An S4 subclass from BinMat class.
}
\details{
The elements of the list are defined as follow.

    \strong{Apriori} a numeric vector. \strong{Scx} a numeric matrix.
    \strong{Scnx} a numeric matrix.
}
\section{Slots}{

\describe{
\item{\code{Score}}{a list object whit 3 elements, Apriori, Scx and Scnx.}
}}
\examples{

}
\author{
Enrique Del Callejo Canal (\email{edelcallejoc@gmail.com}).
}

