% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts.R
\docType{methods}
\name{counts}
\alias{counts}
\alias{counts,BinMat-method}
\alias{counts,BinMat,ANY-method,ANY-method}
\title{Counting matrices for spatial data mining}
\usage{
counts(x.mat, target = NULL, bioclim = NULL)

\S4method{counts}{BinMat}(x.mat, target = NULL, bioclim = NULL)
}
\arguments{
\item{x.mat}{a \code{BinMat} object. See documentation of
\code{\link[rspecies]{id_pts}} when \code{colnames}.}

\item{target}{\code{logical}, \code{character}, \code{numeric} or \code{NULL}.
A vector with names, position (integer) or boolean that especifies the
columns that should be considered as target.}

\item{bioclim}{a \code{BinMat} object. See documentation of
\code{\link[rspecies]{id_pts}} when \code{colnames}.
This matrix represents the cells identification for bioclim
variables (see \code{\link[rspecies]{raster_breaks}}()).}
}
\value{
Returns a \code{\link[rspecies]{BinMatCount} object.}
}
\description{
Take a binary matrix from \code{\link[rspecies]{id_pts}}
    and calculates the counting matrices for spatial data mining.
    The counting matrices are for success/success (1/1), success/failure
    (1/0) and failure/failure (0/0).
}
\details{
The elements of the list are defined as follow.

    \strong{N} an integer. It corresponds to the number of rows in the matrix model
    \code{x.mat}. \strong{Nc} an integer vector. If \code{target = NULL} then
    it calculates \code{crossprod(x.mat, ones)}, where ones is a vector of ones with
    the same rows as \code{x.mat}. If \code{target} is different from \code{NULL}
    then it calculates \code{crossprod(x.mat[,target], ones)}. \strong{Nx} an integer vector.
    If \code{target = NULL} then it calculates \code{crossprod(x.mat, ones)},
    where ones is a vector of ones with the same rows as \code{x.mat} (Notice that
    in this case Nc = Nx). If \code{target} is different from \code{NULL} then
    it calculates \code{crossprod(x.mat[,-target], ones)}. \strong{Ncx} an integer matrix. If
    \code{target = NULL} then it calculates \code{crossprod(x.mat, x.mat)}. If
    \code{target} is different from \code{NULL} then it calculates
    \code{crossprod(x.mat[,target], x.mat[,-target])}.

    If \code{bioclim} is different from \code{NULL} similar calculation as above.
    If \code{target = NULL} the matrices used for calculation are \code{x.mat} and
    \code{bioclim}. If \code{target} is different from \code{NULL} then the matrices
    used for calculation are \code{x.mat[, target]} and \code{bioclim}. After the
    claculation, \code{rbind} and \code{cbind} function are used for proper binding.
}
\examples{
library(sp)
library(rgeos)
data(Mex0)
data(mammals)

# Generating de grid from Mex0 data
Mex0.grd <- grd_build(Mex0)

# Identification points of mammals with colnames especified.
names <- paste("X", 1:nlevels(as.factor(mammals@data$nameID)), sep = "")

x.mat <- id_pts(grd = Mex0.grd, pts = mammals, colnames = names)

# Counting matrices
system.time(count.mat <- counts(x.mat))
getN(count.mat)
head(getNc(count.mat))
head(getNx(count.mat))
head(getNcx(count.mat))


# including target

system.time(count.mat <- counts(x.mat, target = c("X1","X2")))
getN(count.mat)
head(getNc(count.mat))
head(getNx(count.mat))
head(getNcx(count.mat))

# including target and bioclim
library(raster)
bioclim<-getData('worldclim', var='bio', res=2.5)

# Extracting values from 19 bioclim variables
bio.sp <- raster_breaks(bioclim, Mex0.grd)
bnames <- as.character(levels(as.factor(bio.sp@data$nameID)))
bio.mat <- id_pts(grd = Mex0.grd, pts = bio.sp, colnames = bnames)

system.time(count.bmat <- counts(x.mat, target = c("X1","X2"), bioclim = bio.mat))
getN(count.bmat)
head(getNc(count.bmat))
head(getNx(count.bmat))
head(getNcx(count.bmat))

}
\seealso{
\code{\link[base]{crossprod}}, \code{\link[base]{rbind}},
    \code{\link[base]{cbind}}, \code{\link[rspecies]{id_pts}},
    \code{\link[rspecies]{grd_build}}, \code{\link[rspecies]{raster_breaks}}.
}
\author{
Enrique Del Callejo Canal (\email{edelcallejoc@gmail.com}).
}
