% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epsilon.R
\docType{methods}
\name{epsilon}
\alias{epsilon}
\alias{epsilon,SpNaBaCounts,SpNaBaProbs-method}
\alias{epsilon,SpNaba-methods,epsilon-methods}
\title{Correlation method for the Spatial Naive Bayes Model}
\usage{
\S4method{epsilon}{SpNaBaCounts,SpNaBaProbs}(x.counts, x.probs)
}
\arguments{
\item{x.counts}{A \code{\link[rspecies]{SpNaBaCounts}} object.}

\item{x.probs}{A \code{\link[rspecies]{SpNaBaProbs}} object.}
}
\value{
\code{epsilon}: An object of class \code{\link[rspecies]{SpNaBaEps}}.
}
\description{
\code{epsilon} is a function to estimates epsilon parameter for
    the Spatial Naive Bayes Model.
}
\examples{
# Using the function id_pts() for spatial data --------

library(sp)
library(rgeos)
library(rgdal)
library(raster)

# Loading data
data(Mex0)
data(mammals)

# Generating de grid from Mex0 data
Mex0.grd<-grd_build(Mex0)

# Identification points of mammals
x.mat<-id_pts(grd = Mex0.grd, pts = mammals)

# Without target declaration --------------------------

system.time(x.counts <- counts(x.mat))

# With target declaration -----------------------------

x.counts <- counts(x.mat, target = 1:10)

# Probability matrices -----------------------------------
x.probs <- probs(x.counts, fac.lap = 0.1)

# Epsilon parameter -----------------------------------

x.eps <- epsilon(x.counts, x.probs)

}
\references{
\url{http://species.conabio.gob.mx/}
}
\seealso{
\code{\link[rspecies]{SpNaBaCounts}}, \code{\link[rspecies]{SpNaBaProbs}},
     \code{\link[rspecies]{SpNaBa-methods}}
}
\author{
Enrique Del Callejo Canal (\email{edelcallejoc@gmail.com}),
    based on implemented algortihms in web platform SPECIES (see References).
}
