% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-methods.R
\docType{methods}
\name{get_FM}
\alias{get_FM}
\alias{get_FM,SpNaBaMatrix-method}
\alias{get_FM,get-methods}
\alias{get_FM,SpNaBaModel-method}
\alias{get_BM}
\alias{get_BM,SpNaBaMatrix-method}
\alias{get_BM,get-methods}
\alias{get_BM,SpNaBaModel-method}
\alias{get_N}
\alias{get_N,SpNaBaCounts-method}
\alias{get_N,get-methods}
\alias{get_N,SpNaBaModel-method}
\alias{get_Nc}
\alias{get_Nc,SpNaBaCounts-method}
\alias{get_Nc,get-methods}
\alias{get_Nc,SpNaBaModel-method}
\alias{get_Nx}
\alias{get_Nx,SpNaBaCounts-method}
\alias{get_Nx,get-methods}
\alias{get_Nx,SpNaBaModel-method}
\alias{get_Ncx}
\alias{get_Ncx,SpNaBaCounts-method}
\alias{get_Ncx,get-methods}
\alias{get_Ncx,SpNaBaModel-method}
\alias{get_Pc}
\alias{get_Pc,SpNaBaProbs-method}
\alias{get_Pc,get-methods}
\alias{get_Pc,SpNaBaModel-method}
\alias{get_Px}
\alias{get_Px,SpNaBaProbs-method}
\alias{get_Px,get-methods}
\alias{get_Px,SpNaBaModel-method}
\alias{get_Pcx}
\alias{get_Pcx,SpNaBaProbs-method}
\alias{get_Pcx,get-methods}
\alias{get_Pcx,SpNaBaModel-method}
\alias{get_Pxc}
\alias{get_Pxc,SpNaBaProbs-method}
\alias{get_Pxc,get-methods}
\alias{get_Pxc,SpNaBaModel-method}
\alias{get_Pxnc}
\alias{get_Pxnc,SpNaBaProbs-method}
\alias{get_Pxnc,get-methods}
\alias{get_Pxnc,SpNaBaModel-method}
\alias{get_Ecx}
\alias{get_Ecx,SpNaBaEps-method}
\alias{get_Ecx,get-methods}
\alias{get_Ecx,SpNaBaModel-method}
\alias{get_Apriori}
\alias{get_Apriori,SpNaBaScore-method}
\alias{get_Apriori,score-methods}
\alias{get_Apriori,SpNaBaModel-method}
\alias{get_Apriori,get-methods}
\alias{get_Scx}
\alias{get_Scx,SpNaBaScore-method}
\alias{get_Scx,get-methods}
\alias{get_Scx,SpNaBaModel-method}
\title{Accesor methods for SpNaBa-objects}
\usage{
get_FM(x.mat)

get_BM(x.mat)

get_N(x.counts)

get_Nc(x.counts)

get_Nx(x.counts)

get_Ncx(x.counts)

get_Pc(x.probs)

get_Px(x.probs)

get_Pcx(x.probs)

get_Pxc(x.probs)

get_Pxnc(x.probs)

get_Ecx(x.eps)

get_Apriori(x.score)

get_Scx(x.score)
}
\arguments{
\item{x.mat}{A \code{\link[rspeciesdev]{SpNaBaMatrix}} or
\code{\link[rspeciesdev]{SpNaBaModel}} object.}

\item{x.counts}{A \code{\link[rspeciesdev]{SpNaBaCounts}} object  or
\code{\link[rspeciesdev]{SpNaBaModel}} object.}

\item{x.probs}{A \code{\link[rspeciesdev]{SpNaBaProbs}} object  or
\code{\link[rspeciesdev]{SpNaBaModel}} object.}

\item{x.eps}{A \code{SpNaBaEps} object, the resulting object of epsilon function
or A \code{SpNaBaModel} object. The resulting object of NBModel function.}

\item{x.score}{A \code{SpNaBaScore} object. The resulting object of score function
or A \code{SpNaBaModel} object. The resulting object of NBModel function.}
}
\value{
\code{get_FM} return the slot \code{freqmatrix} a \code{SpNaBaMatrix} object.

\code{get_BM} return the slot \code{binmatrix} a \code{SpNaBaMatrix} object.

\code{get_N} return the slot \code{N} a \code{SpNaBaCounts} object.

\code{get_Nc} return the slot \code{Nc} a \code{SpNaBaCounts} object.

\code{get_Nx} return the slot \code{Nx} a \code{SpNaBaCounts} object.

\code{get_Ncx} return the slot \code{Ncx} a \code{SpNaBaCounts} object.

\code{get_Pc} return the slot \code{Pc} a \code{SpNaBaProbs} object.

\code{get_Px} return the slot \code{Px} a \code{SpNaBaProbs} object.

\code{get_Pcx} return the slot \code{Pc.x} a \code{SpNaBaProbs} object.

\code{get_Pxc} return the slot \code{Px.c} a \code{SpNaBaProbs} object.

\code{get_Pxnc} return the slot \code{Px.nc} a \code{SpNaBaProbs} object.

\code{get_Ecx} return the slot \code{Ecx} of a \code{SpNaBaEps} object.

\code{get_Apriori} return the slot \code{Apriori} a \code{SpNaBaScore} object.

\code{get_Scx} return the slot \code{Scx} of a \code{SpNaBaScore} object.
}
\description{
\code{get_} functions are methods to extract slot from \code{SpNaBaMatrix},
    \code{SpNaBaCounts}, \code{SpNaBaProbs} objects. The accesor functions are build in
    with a snake style; i.e. get_suffix, where the suffix are de abbreciation of the slot
    to extract. For example, the suffix \code{FM} (\code{get_FM()}) stands to refer the
    \code{freqmatrix} slot of the \code{SpNaBaMatrix} object. In the same way, the suffix
    \code{BM} (\code{get_BM()}) stands to refer the \code{binmatrix} of the
    \code{SpNaBaMatrix} object.
}
\examples{
# Using the function id_pts() for spatial data --------

library(sp)
library(rgeos)
library(rgdal)
library(raster)

# Loading data
data(Mex0)
data(mammals)

# Generating de grid from Mex0 data
Mex0.grd<-grd_build(Mex0)

# Identification points of mammals
x.mat<-id_pts(grd = Mex0.grd, pts = mammals)

# Extracting Frequqency matrix

FM <- get_FM(x.mat)

FM[1:5,1:5]

# Extracting Binary matrix

BM <- get_BM(x.mat)

BM[1:5,1:5]

# Calculating counts

x.counts <- counts(x.mat)

# Extracting N

N <- get_N(x.counts)
N

# Extracting Nc

Nc <- get_Nc(x.counts)
Nc

# Extracting Nx

Nx <- get_Nx(x.counts)
Nx

# Extracting Ncx

Ncx <- get_Ncx(x.counts)
Ncx

# Calculating probabilities

x.probs <- probs(x.counts, fac.lap = 0.1)

# Extracting Pc

Pc <- get_Pc(x.probs)
Pc

# Extracting Px

Px <- get_Px(x.probs)
Px

# Extracting Pc.x

Pc.x <- get_Pcx(x.probs)
Pc.x

# Extracting Px.c

Px.c <- get_Pxc(x.probs)
Px.c

# Extracting Px.nc

Px.nc <- get_Pxnc(x.probs)
Px.nc

# Calculating epsilon

x.eps <- epsilon(x.counts, x.probs)


# Extracting Epsilon matrix

Ecx <- get_Ecx(x.eps)

Ecx[1:5,1:5]

# Calculating score

x.score <- score(x.probs)


# Extracting Apriori slot

S.Apriori <- get_Apriori(x.score)

S.Apriori[1:5]

# Extracting Score matrix

Scx <- get_Scx(x.score)

Scx[1:5, 1:5]
}
\seealso{
\code{\link[rspeciesdev]{SpNaBaMatrix}}
}
\author{
Enrique Del Callejo Canal (\email{edelcallejoc@gmail.com}).
}
