% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-methods.R
\docType{methods}
\name{get}
\alias{get}
\alias{getName_ID}
\alias{getDMNB}
\alias{getBMNB}
\alias{getN}
\alias{getNc}
\alias{getNx}
\alias{getNcx}
\alias{getPc}
\alias{getPx}
\alias{getPcx}
\alias{getPcnx}
\alias{getPxc}
\alias{getPxnc}
\alias{getPnxc}
\alias{getPnxnc}
\alias{getEcx}
\alias{getEncx}
\alias{getApr}
\alias{getScx}
\alias{getScnx}
\title{Accesor methods for BinMat class and subclass}
\usage{
getName_ID(object)

getDMNB(object)

getBMNB(object)

getN(object)

getNc(object)

getNx(object)

getNcx(object)

getPc(object)

getPx(object)

getPcx(object)

getPcnx(object)

getPxc(object)

getPxnc(object)

getPnxc(object)

getPnxnc(object)

getEcx(object)

getEncx(object)

getApr(object)

getScx(object)

getScnx(object)
}
\arguments{
\item{object}{A \code{\link[rspecies]{BinMat}} object.}
}
\value{
return a slot of BinMat object.
}
\description{
The accesor methods for BinMat-class, each function is a
    method to acces to each slot in BinMat-class
}
\author{
Enrique Del Callejo Canal (\email{edelcallejoc@gmail.com}).
}
