% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_pts.R
\docType{methods}
\name{id_pts}
\alias{id_pts}
\title{Identify points coordinates inside each cell of a grid.}
\usage{
id_pts(grd, pts, contain = c("simple", "properly"))
}
\arguments{
\item{grd}{A \code{SpatialPolygonsDataFrame} object from the
\code{grd_build} function.}

\item{pts}{A \code{SpatialPointsDataFrame} object.}

\item{contain}{Character (\code{simple} or \code{properly}).
By default \code{simple}, it implies a simple contention
evaluation (see \code{\link[rgeos]{gContains}})).
\code{properly} suggests a properly contention evaluation
(see \code{\link[rgeos]{gContainsProperly}})).}
}
\value{
An object of class \code{SpatialPolygonsDataFrame}. The data slot
    contains a data.frame with a column "id" and m columns. The \eqn{m}
    columns represent the sample intensity in each cell of the grid for
    each factor.
}
\description{
This function takes a \code{SpatialPointsDataFrame}
    object and compares each point with the boundaries of each
    cell in a grid (\code{SpatialPolygonsDataFrame} from the
    \code{grd_build} function.
}
\details{
\code{id_pts()} extract the number of observations per cell
    for each factor. It uses parallel programing to reduce the computation
    time. This function depends on \code{parallel}, \code{doParallel} and
    \code{foreach} packages. By defaul the number of cores for parallel
    computing is two, but it is posible to extend the number of cores
    by modify the body of the funtion at line \code{cores <- 2}
}
\examples{

library(sp)
library(rgeos)
library(rgdal)
library(raster)

# Loading data
data(Mex0)
data(mammals)

# Generating de grid from Mex0 data
Mex0.grd<-grd_build(Mex0)

# Identification points of mammals
system.time(x.mat<-id_pts(grd = Mex0.grd, pts = mammals))

# Loading extra libraries for ggplot2
library(plyr)
library(ggplot2)
library(mapproj)

Mex0.grd_for <- fortify(Mex0.grd)
Mex0.grd_df <- plyr::join(Mex0.grd_for,
 data.frame(id = rownames(x.mat@freqmatrix ), x.mat@freqmatrix, stringsAsFactors = FALSE), by="id")

ggplot(data = Mex0.grd_df, aes(x = long, y = lat, group = group, fill = Alouatta_palliata)) +
    geom_polygon(colour = "gray") +
    scale_fill_gradient(low = "white", high = "red") +
    coord_quickmap()


}
\references{
\url{http://species.conabio.gob.mx/}
}
\author{
Enrique Del Callejo Canal (\email{edelcallejoc@gmail.com}),
    based on implemented algortihms in web platform SPECIES (see References).
}
