% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NBModel.R
\docType{methods}
\name{NBModel}
\alias{NBModel}
\alias{NBModel,SpatialPointsDataFrame,SpatialPolygonsDataFrame-method}
\alias{NBModel,SpNaBa-methods,NBModel-methods}
\alias{NBModel,SpNaBaMatrix,ANY-method}
\title{Function to fit Spatial Naive Bayes Model}
\usage{
NBModel(points, map, target = NULL, fac.lap = NULL, isgrid = FALSE,
  contain, ...)
}
\arguments{
\item{points}{A \code{\link[sp]{SpatialPointsDataFrame-class}} object or
a \code{\link[rspecies]{SpNaBaMatrix-class}} object.}

\item{map}{A \code{\link[sp]{SpatialPointsDataFrame-class}} object or NULL. If NULL, \code{points}
argument must be a \code{\link[rspecies]{SpNaBaMatrix-class}} object.}

\item{target}{\code{character} or \code{numeric} or \code{logical} or NULL. Indicates the columns
of binmatrix slot to be consired as target factors.}

\item{fac.lap}{\code{numeric} or \code{NULL}. it indicates the laplace factor to be applied.
If \code{NULL} or \code{missing} the laplace factor applied is 0.01.}

\item{isgrid}{logical (\code{FALSE} by default). If is TRUE idicates that the map is gridded,
if is FALSE the grid is calculated based on the bounding box of the map.}

\item{contain}{Character (\code{simple} or \code{properly}).
By default \code{simple}, it implies a simple contention
evaluation (see \code{\link[rgeos:gContains]{rgeos::gContains()}})).
\code{properly} suggests a properly contention evaluation
(see \code{\link[rgeos]{gContainsProperly}})). It requieres that
points argument should be a \code{\link[sp]{SpatialPointsDataFrame-class}} object
and map argument should be declared as a \code{\link[sp]{SpatialPolygonsDataFrame-class}} object.}

\item{...}{Arguments pass to \code{grd_build} function. See \code{\link[rspecies]{grd_build}}.}
}
\value{
\code{NBModel}: An object of class \code{\link[rspecies]{SpNaBaModel}}.
}
\description{
\code{NBModel} is a function to fit Spatial Naive Bayes Model.
}
\examples{
# Using the function id_pts() for spatial data --------

library(sp)
library(rgeos)
library(rgdal)
library(raster)

# Loading data
data(Mex0)
data(mammals)

# Using an SpatialPointsDataFrame and
# SpatialPolygonsDataFrame object

x.model <- NBModel(mammals, Mex0, target = 1:10, fac.lap = 0.01)


# Using an SpNaBaMAtrix object

# Grid
Mex0.grd<-grd_build(Mex0)

# Identification points of mammals
x.mat<-id_pts(grd = Mex0.grd, pts = mammals)

x.model <- NBModel(x.mat, target = 1:10, fac.lap = 0.01)

}
\references{
\url{http://species.conabio.gob.mx/}
}
\seealso{
\code{\link[rspecies]{SpNaBaMatrix-class}}, \code{\link[rspecies]{SpNaBaCounts}},
    \code{\link[rspecies]{SpNaBaProbs}}, \code{\link[rspecies]{SpNaBa-methods}},
    \code{\link[rspecies]{SpNaBaEps}}, \code{\link[rspecies]{epsilon}},
    \code{\link[rspecies]{SpNaBaScore}}, \code{\link[rspecies]{score}}
}
\author{
Enrique Del Callejo Canal (\email{edelcallejoc@gmail.com}),
    based on implemented algortihms in web platform SPECIES (see References).
}
