% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-methods.R
\docType{methods}
\name{plot-methods}
\alias{plot}
\alias{plot-methods}
\title{An S4 method for SpatialPolygonsDataFrame and BinMat classes objects.}
\usage{
\S4method{plot}{SpatialPolygonsDataFrame,BinMat}(x, y, target = NULL,
  sou_den = TRUE, leaflet = FALSE, ...)

\S4method{plot}{SpatialPolygonsDataFrame,BinMatPred}(x, y, target = NULL,
  leaflet = FALSE, ...)
}
\arguments{
\item{x}{A \code{SpatialPolygonsDataFrame} object from
\code{\link[rspecies]{grd_build}} function.}

\item{y}{A \code{\linkS4class{BinMat}} object from
\code{\link[rspecies]{id_pts}} function.}

\item{target}{\code{logical}, \code{character}, \code{numeric} or \code{NULL}.
A vector with names, position (integer or boolean) that especifies the
columns of slots DMNB or BMNB from \code{\linkS4class{BinMat}}
that should be considered as target.}

\item{leaflet}{logical. If \code{TRUE} the function uses the leaflet package
and methods. If \code{FALSE} the plotting method from sp package are used.}

\item{...}{Parameters passed to other functions (See
\code{\link[graphics]{par}}).}

\item{sou.den}{logical. If \code{TRUE} the color palette is based on DMNB
slot from \code{\linkS4class{BinMat}} object. If \code{FALSE} the color
palette is based in BMNB slot from \code{\linkS4class{BinMat}} object.}
}
\description{

}
\author{
Enrique Del Callejo Canal (\email{edelcallejoc@gmail.com}).
}

