% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_NBmodel.R
\docType{methods}
\name{plotNB}
\alias{plotNB}
\alias{plotNB,SpNaBaModel-method}
\alias{plotNB,SpNaBa-methods,NBModel-methods}
\title{An S4 method for SpNaBaModel class object.}
\usage{
plotNB(nbmodel, ...)
}
\arguments{
\item{nbmodel}{A \code{SpNaBaModel} object from
\code{\link[rspecies]{NBModel}} function.}

\item{...}{Parameters passed to other functions (See
\code{\link[graphics]{par}}).}
}
\description{
A plot methods for class "SpNaBaModel".
}
\details{
This function creates a list with five ggplot objects.

         The first element of the list is called 'Smp_density'.
         It shows an histogram of the density of cell's occurrence
         per specie. In this sense, the x axis of this graph represents
         the number of cells with an occurrence and the y axis shows
         the density of variables (species) in that range of cell's
         occurrence.
}
\examples{
library(sp)
library(rgeos)
data(Mex0)
data(mammals)

# Generating de grid from Mex0 data
Mex0.grd<-grd_build(Mex0)

# Identification points of mammals with colnames = NULL.
x.mat<-id_pts(grd = Mex0.grd, pts = mammals)

# Creando el modelo
x.model <- NBModel(x.mat, target = 1:10, fac.lap = 0.01)

# Graphic report
x.plots <- plotNB(x.model)

}
\author{
Enrique Del Callejo Canal (\email{edelcallejoc@gmail.com}).
}
