% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_predict.R
\docType{methods}
\name{plot_predict}
\alias{plot_predict}
\alias{plot_predict,SpatialPolygonsDataFrame,SpNaBaPred-method}
\alias{plot_predict,SpNaBa-methods,NBModel-methods}
\title{An S4 method for SpatialPolygonsDataFrame and SpNaBaPred classes objects.}
\usage{
plot_predict(map, prediction, target = NULL, ...)
}
\arguments{
\item{map}{A \code{SpatialPolygonsDataFrame} object from
\code{\link[rspecies]{grd_build}} function.}

\item{prediction}{A \code{\linkS4class{SpNaBaPred}} object from
\code{\link[rspecies]{predict}} function.}

\item{target}{\code{logical}, \code{character}, \code{numeric} or \code{NULL}.
A vector with names, position (integer or boolean) that especifies the
columns of the target variables. If length = 1, the prediction map for the
target is displayed. If length > 1, the prediction richness map of the targets
is displayed.}

\item{...}{Parameters passed to other functions (See
\code{\link[graphics]{par}}).}
}
\value{
a ggplot object
}
\description{
A plot methods for classes "SpatialPolygonsDataFrame" and "SpNaBaPred".
}
\examples{
library(sp)
library(rgeos)
data(Mex0)
data(mammals)

# Generating de grid from Mex0 data
Mex0.grd<-grd_build(Mex0)

# Identification points of mammals with colnames = NULL.
x.mat<-id_pts(grd = Mex0.grd, pts = mammals)

# Creando el modelo
x.model <- NBModel(x.mat, target = 1:10, fac.lap = 0.01)

# Calculando la predicción
x.prediction <- predict(x.model, apr_inc = FALSE)

# Plot of one target
plot_predict(Mex0.grd, x.prediction, target = 1)

# plot prediction of richness
plot_predict(Mex0.grd, x.prediction, target = 1:10)


}
\author{
Enrique Del Callejo Canal (\email{edelcallejoc@gmail.com}).
}
