% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\docType{methods}
\name{predict}
\alias{predict}
\alias{predict,SpNaBaModel,logical-method}
\alias{predict,SpNaBa-methods,NBModel-methods}
\title{Predict function for Naive Bayes Spatial Model}
\usage{
predict(object, apr_inc, ...)
}
\arguments{
\item{object}{A \code{SpNaBaModel} object from
\code{\link[rspeciesdev]{NBModel}} function.}

\item{apr_inc}{logical. If TRUE the apriori score is included for
prediction calculation.}

\item{...}{pass to other methods.}
}
\description{
This function allows to calculate the classification of each cell in the
    GRID, based on the score function of the naive Bayes model.
    See \code{\link[rspeciesdev]{score}} function.
}
\examples{
library(sp)
library(rgeos)
data(Mex0)
data(mammals)

# Generating de grid from Mex0 data
Mex0.grd<-grd_build(Mex0)

# Identification points of mammals with colnames = NULL.
x.mat<-id_pts(grd = Mex0.grd, pts = mammals)

x.model <- NBModel(x.mat, target = 1:10, fac.lap = 0.01)

x.prediction <- predict(x.model, apr_inc = FALSE)

}
\author{
Enrique Del Callejo-Canal (\email{edelcallejoc@gmail.com}).
}
