% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probs.R
\docType{methods}
\name{probs}
\alias{probs}
\title{Probability matrices for spatial data mining.}
\usage{
\S4method{probs}{BinMatCount}(count.mat, lap_fac = 0.1)
}
\arguments{
\item{count.mat}{A BinMatCount object. See documentation
\code{\link[rspecies]{BinMatCount}}.}

\item{lap_fac}{numeric. See \code{\link[rspecies]{laplace}}.}
}
\value{
This function returns a \code{BinMatEps} object.
}
\description{
Take a \code{list} object from \code{\link[rspecies]{counts}}
    and calculates the probability matrices for spatial data mining.
    The probability matrices are for joint and conditional probabilities.
}
\details{
The elements of the list are defined as follow.

    \strong{Pc} a numeric vector. The probability of success of the target.
    \strong{Px} a numeric vector. The probability of success of the explanatory
    variables. \strong{Pcx} a numeric matrix. The conditional
    probabilies \eqn{P(C|X)}. Rows correspond target C and columns correspond
    the explanatory variables X. \strong{Pcnx}a numeric matrix. The conditional
    probabilies \eqn{P(C|X')}, where \eqn{X'} is the complement of X.
    Rows correspond target C and columns correspond  the explanatory variables X'.
    \strong{Pxc} a numeric matrix. The conditional probabilies \eqn{P(X|C)}. Rows
    correspond  the explanatory variables X and columns correspond target C.
    \strong{Pxnc} a numeric matrix. The conditional probabilies \eqn{P(X|C')}, where
    C' is the complement of C. Rows correspond  the explanatory variables X and
    columns correspond C'. \strong{Pnxc} a numeric matrix. The conditional probabilies
    \eqn{P(X'|C)}. Rows correspond  the explanatory variables X' and columns correspond
    C'. \strong{Pnxnc} a numeric matrix. The conditional probabilies \eqn{P(X'|C')}.
    Rows correspond  the explanatory variables X' and columns correspond C'.
}
\examples{
library(sp)
library(rgeos)
load('./data/Mex0.rda')
load('./data/mammals.RData')

# Generating de grid from Mex0 data
Mex0.grd<-grd_build(Mex0)

# Identification points of mammals with colnames = NULL.
x.mat<-id_pts(grd = Mex0.grd, pts = mammals, colnames = NULL)

# Counting matrices
count.mat<-counts(x.mat)

# Probability matrices
system.time(prob.mat<-probs(count.mat, lap_fac = 0.1))
}
\author{
Enrique Del Callejo Canal (\email{edelcallejoc@gmail.com}).
}

