% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set-BinMatCount.R
\name{set-BinMatCount}
\alias{set-BinMatCount}
\alias{set-BinMatCount}
\alias{setN,BinMatCount}
\alias{set-BinMatCount}
\alias{setNc,BinMatCount}
\alias{set-BinMatCount}
\alias{setNx,BinMatCount}
\alias{set-BinMatCount}
\alias{setNcx,BinMatCount}
\title{Replace methods for BinMat class and subclass}
\usage{
setN(object) <- value

setNc(object) <- value

setNx(object) <- value

setNcx(object) <- value
}
\arguments{
\item{object}{A \code{\link[rspecies]{BinMatCount}} object.}

\item{value}{An object for replacement the slots of BinMatCount.
'setN' an integer vector of length 1. 'setNc' an integer vector.
'setNx' an integer vector. 'setNcx' an integer matrix. (See
\code{\link[rspecies]{BinMatCount}} for more details.)}
}
\value{
return a slot of BinMat object.
}
\description{
The setters methods for BinMatCount-class, each function is a
    method to acces to each slot in BinMat-class
}
\author{
Enrique Del Callejo Canal (\email{edelcallejoc@gmail.com}).
}
