% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set-methods.R
\docType{methods}
\name{set}
\alias{set}
\alias{setN}
\alias{setN-gen}
\alias{setNc}
\alias{setNc-gen}
\alias{setNcx}
\alias{setNcx-gen}
\alias{setNx}
\alias{setNx-gen}
\title{Replace methods for BinMat class and subclass}
\usage{
setN(object) <- value

\S4method{setN}{BinMatCount}(object) <- value

setNc(object) <- value

\S4method{setNc}{BinMatCount}(object) <- value

setNx(object) <- value

\S4method{setNx}{BinMatCount}(object) <- value

setNcx(object) <- value

\S4method{setNcx}{BinMatCount}(object) <- value
}
\arguments{
\item{object}{A \code{\link[rspecies]{BinMatCount}} object.}

\item{value}{An object for replacement the slots of BinMatCount.
'setN' an integer vector of length 1. 'setNc' an integer vector.
'setNx' an integer vector. 'setNcx' an integer matrix. (See
\code{\link[rspecies]{BinMatCount}} for more details.)}
}
\value{
return a slot of BinMat object.
}
\description{
The setters methods for BinMatCount-class, each function is a
    method to acces to each slot in BinMat-class
}
\author{
Enrique Del Callejo Canal (\email{edelcallejoc@gmail.com}).
}

