% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_methods.R
\docType{methods}
\name{smp_stats}
\alias{smp_stats}
\alias{smp_stats,SpNaBaMatrix-method}
\alias{smp_stats,SpNaBaEps-method}
\alias{smp_stats,SpNaBaScore-method}
\alias{smp_stats,SpNaBaModel-method}
\title{Basic statistics to evaluate sample and model parameters from a Naive Bayes model matrix.}
\usage{
smp_stats(x)
}
\arguments{
\item{x}{A \code{\link[rspeciesdev]{SpNaBaMatrix}}, \code{\link[rspeciesdev]{SpNaBaEps}},
\code{\link[rspeciesdev]{SpNaBaScore}} or \code{\link[rspeciesdev]{SpNaBaModel}} object.}
}
\value{
An object of class \code{\link[data.table]{data.table}}.
}
\description{
This function takes a \code{SpNaBaMatrix} or \code{SpNaBaEps}
    object and extract basic statistics about sample and parameters epsilon and score.
}
\details{
\itemize{
      \item{SpNaBaMatrix:}{Return a data.table with the follow structure.}
      \itemize{
        \item{Factor:}{ The name of the factors included in the model.}
        \item{N:}{ Number of blocks in the partition of the sample space. If the model is not
          for spatial data N represents the number of cases in the model. If the model is
          for spatial data N represents the number of cell in the GRID
          (see \code{\link[rspeciesdev]{grd_build}}).}
        \item{Nw:}{ Number of blocks with at least one observation of factor \eqn{w}. If the model
          is not for spatial data Nw represents the number of cases with presences of factor
          \eqn{w}. If the model is for spatial data Nw represents the number of cell in the GRID
          with at least one obervations of the factor \eqn{w}.}
        \item{minObs:}{ the minimum number of observatios per case of factor \eqn{w}. If the model
          is not for spatial data minObs = 1, i.e. one observation of factor \eqn{w} for any case
          with a presence of factor \eqn{w}. If the model is for spatial data minObs represents
          the minimum number of observations of factor \eqn{w} of all cells with at least one
          observations of factor \eqn{w}.}
        \item{medianObs:}{ the median of the number of observatios per case of factor \eqn{w}. If
          the model is not for spatial data medianObs = 1, i.e. one observation of factor \eqn{w}
          for any case with a presence of factor \eqn{w}. If the model is for spatial data
          medianObs represents the median of the number of observations of factor \eqn{w} of all
          cells with at least one observations of factor \eqn{w}.}
        \item{maxObs:}{ the maximum number of observatios per case of factor \eqn{w}. If the model
          is not for spatial data maxObs = 1, i.e. one observation of factor \eqn{w} for any case
          with a presence of factor \eqn{w}. If the model is for spatial data maxObs represents
          the maximum number of observations of factor \eqn{w} of all cells with at least one
          observations of factor \eqn{w}.}
        \item{TotalObs:}{ the total number of observatios of factor \eqn{w}. If the model is not
          for spatial data TotalObs = Nw, i.e. the total number of observations of factor
          \eqn{w} is equal to the number of cases with a presences of factor \eqn{w}. If the model is
          for spatial data TotalObs represents the total number of observations of factor \eqn{w}.}
        }


    \item{SpNaBaEps and SpNaBaScore:}{A List of two data.table with the follow structure.
          The first element of the list maps the epsilon of target variables.
          The second element of the list maps the epsilon of predictor variables.}
      \itemize{
          \item{Factor:}{ The name of the factors included in the model.}
          \item{Mean:}{ The mean of epsilon.}
          \item{Variance:}{ The Variance of epsilon.}
          \item{Std.Dev:}{ The Std.Dev of epsilon.}
          \item{Lower:}{ The 25\% quantile of epsilon.}
          \item{Median:}{ The Median of epsilon.}
          \item{Upper:}{ The 75\% quantile of epsilon.}
          \item{Max:}{ The Max of epsilon.}
        }

      \item{SpNaBaModel:}{A List of 3 dobject. Each one with the descriptions of
          the function applied to the objects described below.
        }
  }
}
\examples{
# Using the function id_pts() for spatial data --------

library(sp)
library(rgeos)
library(rgdal)
library(raster)

# Loading data
data(Mex0)
data(mammals)

# Generating de grid from Mex0 data
Mex0.grd<-grd_build(Mex0)

# Identification points of mammals
x.mat<-id_pts(grd = Mex0.grd, pts = mammals)

smp_stats(x.mat)

# Checking stats for epsilon measure --------

# Counts calculation ----------------------------------

x.counts <- counts(x.mat, target = 1:10)

# Probability calculation -----------------------------

x.probs <- probs(x.counts)

# Estimating epsilon ----------------------------------

x.eps <- epsilon(x.counts, x.probs)

# Statistics

smp_stats(x.eps)

# Checking stats for score measure --------

# Score parameter -----------------------------------

x.score <- score(x.probs)

# Statistics

smp_stats(x.score)

# Fitting model --------

x.model <- NBModel(x.mat, target = 1:10, fac.lap = 0.01)

# Statistics

smp_stats(x.model)

}
\references{
\url{http://species.conabio.gob.mx/}
}
\seealso{
\code{SpNaBaMatrix}, \code{grd_build}, \code{id_pts}

\code{SpNaBaEps}, \code{epsilon}

\code{SpNaBaScore}, \code{score}

\code{SpNaBaModel}, \code{NBModel}
}
\author{
Enrique Del Callejo Canal (\email{edelcallejoc@gmail.com}),
    based on implemented algortihms in web platform SPECIES (see References).
}
