% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpNaBa_methods.R
\docType{methods}
\name{counts}
\alias{counts}
\alias{counts,SpNaBaMatrix-method}
\alias{counts,SpNaBa-methods}
\alias{laplace}
\alias{laplace,SpNaBaCounts-method}
\alias{laplace,SpNaBa-methods}
\alias{probs}
\alias{probs,SpNaBaCounts-method}
\alias{probs,SpNaBa-methods}
\title{Spatial Naive Bayes Model methods}
\usage{
\S4method{counts}{SpNaBaMatrix}(x.mat, target = NULL)

\S4method{laplace}{SpNaBaCounts}(x.counts, fac.lap = NULL)

\S4method{probs}{SpNaBaCounts}(x.counts, fac.lap = NULL)
}
\arguments{
\item{x.mat}{A \code{\link[rspecies]{SpNaBaMatrix}} object.}

\item{target}{\code{character} or \code{numeric} or \code{logical} or NULL. Indicates the columns
of binmatrix slot to be consired as target factors.}

\item{x.counts}{A \code{\link[rspecies]{SpNaBaCounts}} object.}

\item{fac.lap}{\code{numeric} or \code{NULL}. it indicates the laplace factor to be applied.
If \code{NULL} or \code{missing} the laplace factor applied is 0.01.}
}
\value{
\code{counts}: An object of class \code{\link[rspecies]{SpNaBaCounts}}.

\code{laplace}: An object of class \code{\link[rspecies]{SpNaBaCounts}}.

\code{probs}: An object of class \code{\link[rspecies]{SpNaBaProbs}}.
}
\description{
\code{counts} is a function to extract frequency information
    for the Spatial Naive Bayes Model.

\code{laplace} is a function to applied the Laplace smothing method
    to an object of class \code{SpNaBaCounts}.

Take a \code{SpNaBaCounts} object from \code{\link[rspecies]{counts}}
    and calculates the probability matrices for spatial data mining.
    The probability matrices are for joint and conditional probabilities.
}
\examples{
# Using the function id_pts() for spatial data --------

library(sp)
library(rgeos)
library(rgdal)
library(raster)

# Loading data
data(Mex0)
data(mammals)

# Generating de grid from Mex0 data
Mex0.grd<-grd_build(Mex0)

# Identification points of mammals
x.mat<-id_pts(grd = Mex0.grd, pts = mammals)

# Without target declaration --------------------------

system.time(x.counts <- counts(x.mat))

# With target declaration -----------------------------

x.counts <- counts(x.mat, target = 1:10)


# Appliying the Laplace smothing ----------------------

x.counts.ls <- laplace(x.counts, fac.lap = 0.1)


# Probability matrices -----------------------------------
x.probs <- probs(x.counts, fac.lap = 0.1)

}
\references{
\url{http://species.conabio.gob.mx/}
}
\seealso{
\code{\link[rspecies]{SpNaBaCounts}}, \code{SpNaBaMatrix}, \code{grd_build}, \code{id_pts}
}
\author{
Enrique Del Callejo Canal (\email{edelcallejoc@gmail.com}),
    based on implemented algortihms in web platform SPECIES (see References).
}
