% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpNaBaCounts.R
\docType{class}
\name{SpNaBaCounts-class}
\alias{SpNaBaCounts-class}
\alias{initialize,SpNaBaCounts-method}
\alias{SpNaBaCounts}
\alias{show,SpNaBaCounts-method}
\title{Class SpNaBaCounts}
\usage{
SpNaBaCounts(N, Nc, Nx, Ncx, ...)
}
\arguments{
\item{N}{\code{vector} of length 1. See slot section for more details.}

\item{Nc}{\code{vector}. See slot section for more details.}

\item{Nx}{\code{vector}. See slot section for more details.}

\item{Ncx}{\code{matrix}. See slot section for more details.}

\item{...}{argument pass to other methods.}
}
\value{
An object of class \code{SpNaBaCounts}. See slot section for more details.
}
\description{
SpNaBaCounts is used to store frequency information of the Naive Bayes model.
}
\section{Slots}{

\describe{
\item{\code{N}}{\code{numeric} of length 1. It contains the number of blocks of the partition
of the sample space. If the model is not for spatial data, N is the number of cases in the
sample. If the model is for spatial data, N is the number of cells in the grid
(see \code{gr_build}).}

\item{\code{Nc}}{\code{numeric}. Each element represents the number of blocks with at least one
observation of target factor $c$. If the model is not for spatial data Nc
represents the number of cases with presences of target factor $c$. If the
model is for spatial data Nc represents the number of cell in the GRID
with at least one obervations of the target factor $c$.}

\item{\code{Nx}}{\code{numeric}. Each element represents the number of blocks with at least one
observation of predictor factor $x$. If the model is not for spatial data Nx
represents the number of cases with presences of predictor factor $x$. If the
model is for spatial data Nx represents the number of cell in the GRID
with at least one obervations of the predictor factor $x$.}

\item{\code{Ncx}}{\code{matrix}. Each element represents the number of blocks with at least one
observation of target factor $c$ and predictor factor $x$. If the model is not
for spatial data Ncx represents the joint frequency of target factor $c$ and
the predictor factor $x$. If the model is for spatial data Ncx represents the
number of cell in the GRID with at least one obervations of the target factor
$c$ and the predictor factor $x$.}
}}

\examples{

# Using SpNaBaMatrix constructor ----------------------

a <- matrix(rbinom(100, 100, runif(100)), nrow = 10, ncol = 10,
  dimnames = list(paste0("id:", 1:10), letters[1:10]))

x.mat <- SpNaBaMatrix(freqmatrix = a)

SpNaBaCounts(N = nrow(x.mat@binmatrix),
             Nc = colSums(x.mat@binmatrix),
             Nx = colSums(x.mat@binmatrix),
             Ncx = t(x.mat@binmatrix) \%*\% x.mat@binmatrix)


}
\references{
\url{http://species.conabio.gob.mx/}
}
\author{
Enrique Del Callejo Canal (\email{edelcallejoc@gmail.com}),
    based on implemented algortihms in web platform SPECIES (see References).
}
