% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpNaBaMatrix.R
\docType{class}
\name{SpNaBaMatrix-class}
\alias{SpNaBaMatrix-class}
\alias{initialize,SpNaBaMatrix-method}
\alias{SpNaBaMatrix}
\alias{show,SpNaBaMatrix-method}
\title{Class SpNaBaMatrix}
\usage{
SpNaBaMatrix(freqmatrix, ...)
}
\arguments{
\item{freqmatrix}{\code{matrix}. typically, the output of \code{id_pts()} function.
See slots section for more details.}

\item{...}{argument pass to other methods.}
}
\value{
An object of class \code{SpNaBaMatrix}. See slot section for more details.
}
\description{
SpNaBaMatrix is used to fit Spatial Naive Bayes models. It can be used to extract information
    of the sample and to adjust Spatial Naive Bayes models.
}
\section{Slots}{

\describe{
\item{\code{freqmatrix}}{\code{matrix}. A matrix with $N$ rows (rows correspond to the number of blocks
in the GRID) and $M$ columns (columns represent the number of factors included in the model),
where each element $i,j$ shows the number of observations of the factor $j$ in the cell $i$.}

\item{\code{binmatrix}}{\code{matrix}. A matrix with $N$ rows (rows correspond to the number of blocks
in the GRID) and $M$ columns (columns represent the number of factors included in the model),
where each element $i,j$ shows \code{TRUE} if the cell $i$ contains at least one observation
of the factor $j$. \code{FALSE} otherwise.}
}}

\examples{

# Using SpNaBaMatrix constructor ----------------------

a <- matrix(rbinom(100, 100, runif(100)), nrow = 10, ncol = 10,
  dimnames = list(paste0("id:", 1:10), letters[1:10]))

x.mat <- SpNaBaMatrix(freqmatrix = a)

x.mat

# Using the function id_pts() for spatial data --------

library(sp)
library(rgeos)
library(rgdal)
library(raster)

# Loading data
data(Mex0)
data(mammals)

# Generating de grid from Mex0 data
Mex0.grd<-grd_build(Mex0)

# Identification points of mammals
x.mat<-id_pts(grd = Mex0.grd, pts = mammals)

x.mat

}
\references{
\url{http://species.conabio.gob.mx/}
}
\seealso{
\code{\link[rspeciesdev]{grd_build}} and \code{\link[rspeciesdev]{id_pts}}.
}
\author{
Enrique Del Callejo Canal (\email{edelcallejoc@gmail.com}),
    based on implemented algortihms in web platform SPECIES (see References).
}
