% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpNaBaModel.R
\docType{class}
\name{SpNaBaModel-class}
\alias{SpNaBaModel-class}
\alias{initialize,SpNaBaModel-method}
\alias{SpNaBaModel}
\alias{show,SpNaBaModel-method}
\title{Class SpNaBaModel}
\usage{
SpNaBaModel(ModMatrix, Counts, Probs, Epsilon, Score, ...)
}
\arguments{
\item{ModMatrix}{a SpNaBaMatrix. See \code{SpNaBaMatrix}.}

\item{Counts}{a SpNaBaCounts. See \code{SpNaBaCounts}.}

\item{Probs}{a SpNaBaProbs. See \code{SpNaBaProbs}.}

\item{Epsilon}{a SpNaBaEps. See \code{SpNaBaEps}.}

\item{Score}{a SpNaBaScore. See \code{SpNaBaScore}.}

\item{...}{argument pass to other methods}
}
\value{
An object of class \code{SpNaBaModel}. See slot section for more details.
}
\description{
SpNaBaModel is used to store information about the naive Bayes model.
}
\section{Slots}{

\describe{
\item{\code{ModMatrix}}{a SpNaBaMatrix. See \code{SpNaBaMatrix}.}

\item{\code{Counts}}{a SpNaBaCounts. See \code{SpNaBaCounts}.}

\item{\code{Probs}}{a SpNaBaProbs. See \code{SpNaBaProbs}.}

\item{\code{Epsilon}}{a SpNaBaEps. See \code{SpNaBaEps}.}

\item{\code{Score}}{a SpNaBaScore. See \code{SpNaBaScore}.}
}}

\examples{

# Using SpNaBaMatrix constructor ----------------------

a <- matrix(rbinom(100, 100, runif(100, 0, 0.1)), nrow = 10, ncol = 10,
  dimnames = list(paste0("id:", 1:10), letters[1:10]))

x.mat <- SpNaBaMatrix(freqmatrix = a)

# Counts calculation ----------------------------------

x.counts <- counts(x.mat)

# Probability calculation -----------------------------

x.probs <- probs(x.counts)

# Estimating epsilon ----------------------------------

x.eps <- epsilon(x.counts, x.probs)

# Estimating Score

x.score <- score(x.probs)

x.model <- SpNaBaModel(x.mat, x.counts, x.probs,
                       x.eps, x.score)

}
\references{
\url{http://species.conabio.gob.mx/}
}
\author{
Enrique Del Callejo Canal (\email{edelcallejoc@gmail.com}),
    based on implemented algortihms in web platform SPECIES (see References).
}
