% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpNaBaProbs.R
\docType{class}
\name{SpNaBaProbs-class}
\alias{SpNaBaProbs-class}
\alias{initialize,SpNaBaProbs-method}
\alias{SpNaBaProbs}
\alias{show,SpNaBaProbs-method}
\title{Class SpNaBaProbs}
\usage{
SpNaBaProbs(Pc, Px, Pc.x, Px.c, Px.nc, ...)
}
\arguments{
\item{Pc}{numeric \code{vector}. See slot section for more details.}

\item{Px}{numeric \code{vector}. See slot section for more details.}

\item{Pc.x}{numeric \code{matrix}. See slot section for more details.}

\item{Px.c}{numeric \code{matrix}. See slot section for more details.}

\item{Px.nc}{numeric \code{matrix}. See slot section for more details.}

\item{...}{argument pass to other methods.}
}
\value{
An object of class \code{SpNaBaProbs}. See slot section for more details.
}
\description{
SpNaBaProbs is used to store probability information of the Naive Bayes model.
}
\section{Slots}{

\describe{
\item{\code{Pc}}{numeric \code{vector}. It contains the marginal probabilities of the
target variables declared in the NaBa model.}

\item{\code{Px}}{numeric \code{vector}. It contains the marginal probabilities of the
predictor variables declared in the NaBa model.}

\item{\code{Pc.x}}{numeric \code{matrix}. Each element represents the condicional
probability \eqn{P(c|x)}. The sufix \code{c.x} implies that the rows of this
matrix represents the target variables (c) and the columns represents the
predictors variables (x) declared in the Naive Bayes model.}

\item{\code{Px.c}}{numeric \code{matrix}. Each element represents the condicional
probability \eqn{P(x|c)}. The sufix \code{x.c} implies that the rows of this
matrix represents the predictors variables (x) and the columns represents
the target variables (c) declared in the Naive Bayes model.}

\item{\code{Px.nc}}{numeric \code{matrix}. Each element represents the condicional
probability \eqn{P(x|c')}. The sufix \code{x.nc} implies that the rows of this
matrix represents the predictors variables (x) and the columns represents
the complement of the target variables (\eqn{c'}) declared in the Naive Bayes
 model.}
}}

\examples{

# Using SpNaBaMatrix constructor ----------------------

a <- matrix(rbinom(100, 100, runif(100, 0, 0.1)), nrow = 10, ncol = 10,
  dimnames = list(paste0("id:", 1:10), letters[1:10]))

x.mat <- SpNaBaMatrix(freqmatrix = a)

# Counts calculation ----------------------------------

x.counts <- counts(x.mat)

# Probability calculation -----------------------------

x.probs <- probs(x.counts)

}
\references{
\url{http://species.conabio.gob.mx/}
}
\author{
Enrique Del Callejo Canal (\email{edelcallejoc@gmail.com}),
    based on implemented algortihms in web platform SPECIES (see References).
}
