% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BinMatCount-class.R
\docType{methods}
\name{BinMatCount-class}
\alias{BinMatCount-class}
\alias{new-BinMatCount}
\alias{.BinMatCount}
\alias{BinMatCount-Contstructor}
\alias{BinMatCount}
\alias{show-BinMatCount}
\alias{print-BinMatCount}
\title{An S4 subclass from BinMat class.}
\usage{
BinMatCount(name_ID, DMNB, BMNB, Count, ...)

\S4method{show}{BinMatCount}(object)

\S4method{print}{BinMatCount}(x, ...)
}
\arguments{
\item{...}{pass to another methods.}

\item{object}{an object of class BinMatCount.}

\item{x}{an object of class BinMatCount.}
}
\description{
This is a class associated to count methods applied to
    BinMat objects (See \code{\link[rspecies]{counts}}).
}
\details{
The elements of the list are defined as follow.

    \strong{N} an integer. It corresponds to the number of rows in the slot
    name_ID of BinMat object.

    \strong{Nc} a numeric vector.

    \strong{Nx} a numeric vector.

    \strong{Ncx} a numeric matrix.
}
\section{Slots}{

\describe{
\item{\code{name_ID}}{a data.frame. rownames(name_ID) = colnames(DMNB) = colnames(BMNB).
A data containing the ID's for column names in slots DMNB and BMNB.}

\item{\code{DMNB}}{a matrix of integers. Contains the number of observations per cells
per column. The number of columns must be identical to the number of rows
in name_ID.}

\item{\code{BMNB}}{a logical matrix. Contains TRUE for all DMNB > 0, FALSE otherwise.
The number of columns must be identical to the number of rows in name_ID.}

\item{\code{Count}}{a list object whit 4 elements, N, Nc, Nx and Ncx. See details}
}}

\examples{
library(sp)
library(rgeos)
data(Mex0)
data(mammals)

# Generating de grid from Mex0 data
Mex0.grd <- grd_build(Mex0)

# Identification points of mammals with colnames especified.
names <- paste("X", 1:nlevels(as.factor(mammals@data$nameID)), sep = "")

x.mat <- id_pts(grd = Mex0.grd, pts = mammals, colnames = names)

# Counting matrices
system.time(count.mat <- counts(x.mat))
getN(count.mat)
head(getNc(count.mat))
head(getNx(count.mat))
head(getNcx(count.mat))
}
\author{
Enrique Del Callejo Canal (\email{edelcallejoc@gmail.com}).
}
