% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BinMatEps-class.R
\docType{methods}
\name{BinMatEps-class}
\alias{.BinMatEps}
\alias{BinMatEps}
\alias{BinMatEps-Contstructor}
\alias{BinMatEps-class}
\alias{new-BinMatEps}
\alias{print-BinMatEps}
\alias{show-BinMatEps}
\title{An S4 subclass from BinMat class.}
\usage{
BinMatEps(name_ID, DMNB, BMNB, Epsilon, ...)

\S4method{show}{BinMatEps}(object)

\S4method{print}{BinMatEps}(x, ...)
}
\description{
An S4 subclass from BinMat class.
}
\details{
The elements of the list are defined as follow.

    \strong{Ecx}: numeric matrix. With s rows and m columns

    \strong{Pcnx}a numeric matrix. With s rows and m columns.
}
\section{Slots}{

\describe{
\item{\code{Epsilon}}{a list object whit 2 elements, Ecx and Encx.}
}}
\examples{
library(sp)
library(rgeos)
data(Mex0)
data(mammals)

# Generating de grid from Mex0 data
Mex0.grd<-grd_build(Mex0)

# Identification points of mammals with colnames = NULL.
x.mat<-id_pts(grd = Mex0.grd, pts = mammals, colnames = NULL)

# Counting matrices
count.mat<-counts(x.mat)

# Probability matrices
prob.mat<-probs(count.mat, laplace = 0.1)

# Epsilon function
epsilon.mat<-epsilon(prob.mat, count.mat)

}
\author{
Enrique Del Callejo Canal (\email{edelcallejoc@gmail.com}).
}

