% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BinMatEps-class.R
\docType{methods}
\name{BinMatEps-class}
\alias{BinMatEps-class}
\alias{new-BinMatEps}
\alias{.BinMatEps}
\alias{BinMatEps-Contstructor}
\alias{BinMatEps}
\alias{show-BinMatEps}
\alias{print-BinMatEps}
\title{An S4 subclass from BinMat class.}
\usage{
BinMatEps(name_ID, DMNB, BMNB, Epsilon, ...)

\S4method{show}{BinMatEps}(object)

\S4method{print}{BinMatEps}(x, ...)
}
\arguments{
\item{...}{pass to another methods.}

\item{object}{an object of class BinMatEps.}

\item{x}{an object of class BinMatEps.}
}
\description{
This is a class associated to epsilon methods applied to
    BinMatCount and BinMatProb objects (See \code{\link[rspecies]{epsilon}}).
}
\details{
The elements of the list are defined as follow.

    \strong{Ecx}: numeric matrix. With s rows and m columns

    \strong{Pcnx}a numeric matrix. With s rows and m columns.
}
\section{Slots}{

\describe{
\item{\code{name_ID}}{a data.frame. rownames(name_ID) = colnames(DMNB) = colnames(BMNB).
A data containing the ID's for column names in slots DMNB and BMNB.}

\item{\code{DMNB}}{a matrix of integers. Contains the number of observations per cells
per column. The number of columns must be identical to the number of rows
in name_ID.}

\item{\code{BMNB}}{a logical matrix. Contains TRUE for all DMNB > 0, FALSE otherwise.
The number of columns must be identical to the number of rows in name_ID.}

\item{\code{Epsilon}}{a list object whit 2 elements, Ecx and Encx.}
}}

\examples{
library(sp)
library(rgeos)
data(Mex0)
data(mammals)

# Generating de grid from Mex0 data
Mex0.grd<-grd_build(Mex0)

# Identification points of mammals with colnames = NULL.
x.mat<-id_pts(grd = Mex0.grd, pts = mammals, colnames = NULL)

# Counting matrices
count.mat<-counts(x.mat)

# Probability matrices
prob.mat<-probs(count.mat, lap_fac = 0.1)

# Epsilon function
epsilon.mat<-epsilon(prob.mat, count.mat)

}
\author{
Enrique Del Callejo Canal (\email{edelcallejoc@gmail.com}).
}
