% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BinMatPred-class.R
\docType{methods}
\name{BinMatPred-class}
\alias{BinMatPred-class}
\alias{new-BinMatPred}
\alias{.BinMatPred}
\alias{BinMatPred-Constructor}
\alias{BinMatPred}
\alias{show-BinMatPred}
\alias{print-BinMatPred}
\title{An S4 subclass from BinMat class.}
\usage{
BinMatPred(name_ID, DMNB, BMNB, Prediction, ...)

\S4method{show}{BinMatPred}(object)

\S4method{print}{BinMatPred}(x, ...)
}
\arguments{
\item{...}{pass to another methods.}

\item{object}{an object of class BinMatPred.}

\item{x}{an object of class BinMatPred.}
}
\description{
This is a class associated to prediction methods applied to
    BinMatScore objects (See \code{\link[rspecies]{predict}}).
}
\details{
The elements of the list are defined as follow.
}
\section{Slots}{

\describe{
\item{\code{name_ID}}{a data.frame. rownames(name_ID) = colnames(DMNB) = colnames(BMNB).
A data containing the ID's for column names in slots DMNB and BMNB.}

\item{\code{DMNB}}{a matrix of integers. Contains the number of observations per cells
per column. The number of columns must be identical to the number of rows
in name_ID.}

\item{\code{BMNB}}{a logical matrix. Contains TRUE for all DMNB > 0, FALSE otherwise.
The number of columns must be identical to the number of rows in name_ID.}

\item{\code{Prediction}}{a matrix object.}
}}

\examples{
library(sp)
library(rgeos)
data(Mex0)
data(mammals)

# Generating de grid from Mex0 data
Mex0.grd<-grd_build(Mex0)

# Identification points of mammals with colnames = NULL.
x.mat<-id_pts(grd = Mex0.grd, pts = mammals, colnames = NULL)

# Counting matrices
count.mat<-counts(x.mat,  target = c("F.169", "F.272"))

# Probability matrices
prob.mat<-probs(count.mat, lap_fac = 0.1)

#score function
score.mat<-score(prob.mat, count.mat)

# Prediction
pred.mat<-predict(score.mat, apr_inc = FALSE, comp_inc = FALSE)

print(pred.mat)

}
\author{
Enrique Del Callejo Canal (\email{edelcallejoc@gmail.com}).
}
