% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BinMatPred-class.R
\docType{methods}
\name{BinMatPred-class}
\alias{BinMatPred}
\alias{BinMatPred-class}
\alias{Contstructor}
\alias{print}
\title{An S4 subclass from BinMat class.}
\usage{
BinMatPred(name_ID, DMNB, BMNB, Prediction, ...)

\S4method{print}{BinMatPred}(x, ...)
}
\description{
An S4 subclass from BinMat class.
}
\details{
The elements of the list are defined as follow.
}
\section{Slots}{

\describe{
\item{\code{Prediction}}{a matrix object.}
}}
\examples{

}
\author{
Enrique Del Callejo Canal (\email{edelcallejoc@gmail.com}).
}

