% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BinMatProb-class.R
\docType{methods}
\name{BinMatProb-class}
\alias{.BinMatProb}
\alias{BinMatProb}
\alias{BinMatProb-Contstructor}
\alias{BinMatProb-class}
\alias{new-BinMatProb}
\alias{print-BinMatProb}
\alias{show-BinMatProb}
\title{An S4 subclass from BinMat class.}
\usage{
BinMatProb(name_ID, DMNB, BMNB, Prob, ...)

\S4method{show}{BinMatProb}(object)

\S4method{print}{BinMatProb}(x, ...)
}
\description{
An S4 subclass from BinMat class.
}
\details{
The elements of the list are defined as follow.

    \strong{Pc} a numeric matrix. With s rows and 1 column.

    \strong{Px} a numeric matrix. With m rows and 1 column.

    \strong{Pcx} a numeric matrix. With s rows and m columns.

    \strong{Pcnx}a numeric matrix. With s rows and m columns.

    \strong{Pxc} a numeric matrix. With m rows and s columns.

    \strong{Pxnc} a numeric matrix. With m rows and s columns.

    \strong{Pnxc} a numeric matrix. With m rows and s columns.

    \strong{Pnxnc} a numeric matrix. With m rows and s columns.
}
\section{Slots}{

\describe{
\item{\code{Prob}}{a list object whit 8 elements, Pc, Px, Pcx, Pcnx, Pxc, Pxnc,
Pnxc and Pnxnc. See details.}
}}
\examples{
#' #' library(sp)
library(rgeos)
data(Mex0)
data(mammals)

# Generating de grid from Mex0 data
Mex0.grd<-grd_build(Mex0)

# Identification points of mammals with colnames = NULL.
x.mat<-id_pts(grd = Mex0.grd, pts = mammals, colnames = NULL)

# Counting matrices
count.mat<-counts(x.mat,  target = c("F.169", "F.272"))

# Probability matrices
prob.mat<-probs(count.mat, lap_fac = 0.1)

print(prob.mat)
}
\author{
Enrique Del Callejo Canal (\email{edelcallejoc@gmail.com}).
}

