% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BinMatScore-class.R
\docType{methods}
\name{BinMatScore-class}
\alias{.BinMatScore}
\alias{BinMatScore}
\alias{BinMatScore-Contstructor}
\alias{BinMatScore-class}
\alias{new-BinMatScore}
\alias{print-BinMatScore}
\alias{show-BinMatScore}
\title{An S4 subclass from BinMat class.}
\usage{
BinMatScore(name_ID, DMNB, BMNB, Score, ...)

\S4method{show}{BinMatScore}(object)

\S4method{print}{BinMatScore}(x, ...)
}
\arguments{
\item{...}{pass to another methods.}

\item{object}{an object of class BinMatScore.}

\item{object}{an object of class BinMatScore.}
}
\description{
This is a class associated to score method applied to
    BinMatProb objects (See \code{\link[rspecies]{score}}).
}
\details{
The elements of the list are defined as follow.

    \strong{Apriori} a numeric vector. \strong{Scx} a numeric matrix.
    \strong{Scnx} a numeric matrix.
}
\section{Slots}{

\describe{
\item{\code{name_ID}}{a data.frame. rownames(name_ID) = colnames(DMNB) = colnames(BMNB).
A data containing the ID's for column names in slots DMNB and BMNB.}

\item{\code{DMNB}}{a matrix of integers. Contains the number of observations per cells
per column. The number of columns must be identical to the number of rows
in name_ID.}

\item{\code{BMNB}}{a logical matrix. Contains TRUE for all DMNB > 0, FALSE otherwise.
The number of columns must be identical to the number of rows in name_ID.}

\item{\code{Score}}{a list object whit 3 elements, Apriori, Scx and Scnx.}
}}
\examples{

#' library(sp)
library(rgeos)
data(Mex0)
data(mammals)

# Generating de grid from Mex0 data
Mex0.grd<-grd_build(Mex0)

# Identification points of mammals with colnames = NULL.
x.mat<-id_pts(grd = Mex0.grd, pts = mammals, colnames = NULL)

# Counting matrices
count.mat<-counts(x.mat,  target = c("F.169", "F.272"))

# Probability matrices
prob.mat<-probs(count.mat, lap_fac = 0.1)

#score function
score.mat<-score(prob.mat, count.mat)

}
\author{
Enrique Del Callejo Canal (\email{edelcallejoc@gmail.com}).
}

