% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_species.R
\name{get_species}
\alias{get_species}
\alias{get_species_names}
\title{Get data from SPECIE framework}
\usage{
get_species(species = NULL, from = "SPECIES", date = TRUE)

get_species_names(level = "genus", name = NULL, limit = NULL)
}
\arguments{
\item{species}{a character vector of length 1. The name of the specie desired as
web SPECIES platform shows.}

\item{from}{The name of web platform to consult data. At this point
it only works for \url{http://species.conabio.gob.mx}.}

\item{date}{logical. If TRUE the collection date is added.}

\item{level}{a character vector of length 1. The name of the taxonomic rank.
The allowed values are: "kingdom", "phylum", "class", "order", "family",
"genus" and "specie". By default "genus".}

\item{name}{a character vector of length 1. The name of the specie desired as
web SPECIES platform shows.}

\item{limit}{a integer vector of length 1. The first "limit" elements to get from
web SPECIES data base. By default is \code{NULL} and extract all data that
corresponds to the name of the taxonomic rank.}
}
\description{
a function to extract data from SPECIES WEB framework.
}
\note{
\code{get_species}. Only one species per query is allowed. This restriction
    is for memory care purposes. You can use the 'apply' function family to extend
    your search, but it is recommended to do the search one by one to not exceed
    the memory limits.

\code{get_species_names}. The data can be very large so proceed with carefull.
}
\examples{

library(httr)
library(jsonlite)

# get_species() examples

# Lynx rufus
LR<-get_species(species = "Lynx rufus")
LR

# Aedes aegypti
AA<-get_species(species = "Aedes aegypti")
AA


# get_species_names() examples

# Initial capital letter
get_species_names(name = "Aedes")

# lowercase is allowed
get_species_names(name = "aedes")


}
\author{
Enrique Del Callejo Canal (\email{edelcallejoc@gmail.com}).
}

