% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_species.R
\name{get_species}
\alias{get_species}
\title{Get data from SPECIE framework}
\usage{
get_species(species = NULL, from = "SPECIES", date = TRUE)
}
\arguments{
\item{species}{a character vector. The name of the specie desired as
web SPECIES platform shows.}

\item{from}{The name of web platform to consult data.}

\item{date}{logical. If TRUE the recolection date is added.}
}
\description{
a function to extract data from SPECIES WEB framework.
}

