% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_species.R
\name{get_species_}
\alias{get_species_}
\alias{get_species_coords}
\alias{get_species_names}
\alias{get_species_grid}
\alias{get_species_georel}
\title{Get data from SPECIE framework}
\usage{
get_species_coords(species = NULL, from = "SPECIES", date = TRUE)

get_species_names(level = "genus", name = NULL, sublevel, id = TRUE)

get_species_grid(grid_res = 16)

get_species_georel(target, start_level, value, grid_res = 16,
  validation = FALSE, fossil = TRUE, no_date = TRUE, min_occ = 1,
  bioclim = FALSE)
}
\arguments{
\item{species}{a character vector of length 1. The name of the specie desired as
web SPECIES platform shows.}

\item{from}{The name of web platform to consult data. At this point
it only works for \url{http://species.conabio.gob.mx}.}

\item{date}{logical. If TRUE the collection date is added.}

\item{level}{a character vector of length 1. The name of the taxonomic rank.
The allowed values are: "kingdom", "phylum", "class", "order", "family",
"genus" and "specie". By default "genus".}

\item{name}{a character vector of length 1. The name of the specie desired as
web SPECIES platform shows.}

\item{sublevel}{character or NULL. A vector of length 1. The name of taxonomic
rank sublevel to search names. See examples.}

\item{id}{Logical. If \code{TRUE}, id metada is included in the output. If
\code{FALSE} only the names of species are display.}

\item{grid_res}{The resolution of the grid in km. The resolution at this momment are 8, 16, 32, and 64 km.}

\item{target}{a character vector of length 1. The name of the specie desired as
web SPECIES platform shows.}

\item{start_level}{a character vector of length 1. The name of the taxonomic rank.
The allowed values are: "kingdom", "phylum", "class", "order", "family",
"genus" and "specie".}

\item{value}{a character vector of length 1. The taxonomic name associated to
the taxonomic rank (level argument) as web SPECIES platform shows.}

\item{validation}{logical. If TRUE validation process is applied to the model.}

\item{fossil}{logical. If TRUE fossil data is included in the model.}

\item{no_date}{logical. If TRUE data without recolection date is included.}

\item{min_occ}{numeric. The minimum number of cells occurrences (nj).}

\item{bioclim}{logical. If TRUE bioclim data is included in the model.}
}
\value{
\code{get_species_coords()} returns a \code{data.frame} object.

\code{get_species_names()} returns a \code{data.frame} object.

\code{get_species_grid()} returns a \code{SpatialPolygonsDataFrame} object. It includes all the polygons
    that form the grid. Data slot contains the identifiers for each cells.

a data.frame object.
}
\description{
a family functions to extract data from SPECIES WEB framework.
}
\note{
\code{get_species_coords}. Only one species per query is allowed. This restriction
    is for memory care purposes. You can use the 'apply' function family to extend
    your search, but it is recommended to do the search one by one to not exceed
    the memory limits.

\code{get_species_names}. The data can be very large so proceed with carefull.
}
\examples{

library(httr)
library(jsonlite)

# get_species_coords() examples

# Lynx rufus
LR<-get_species_coords(species = "Lynx rufus")
LR

# Aedes aegypti
AA<-get_species_coords(species = "Aedes aegypti")
AA


# get_species_names() examples

# Initial capital letter
get_species_names(name = "Aedes")

# lowercase is allowed
get_species_names(name = "aedes")

# Cheking names of a level
# Cheking class' names of Craniata phylum
get_species_names(level = "phylum", name = "Craniata",
 sublevel = "class", id = FALSE)

# Cheking class' names of Animalia kingdom
get_species_names(level = "kingdom", name = "Animalia",
 sublevel = "class", id = FALSE)



# get_species_grid() examples

# By default resolution is 16 km.
system.time(a16 <- get_species_grid())

system.time(a32 <- get_species_grid(grid_res = 32))



Model1 <- get_species_georel(target = "Aedes aegypti",
    start_level = "class", value = "aves", min_occ = 5)

head(Model1)

}
\references{
\url{http://species.conabio.gob.mx/}
}
\author{
Enrique Del Callejo Canal (\email{edelcallejoc@gmail.com}).
}
