% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_pts.R
\docType{methods}
\name{id_pts}
\alias{id_pts}
\title{Identify points coordinates inside each cell of the grid.}
\usage{
id_pts(grd, pts, contain = c("simple", "properly"), colnames = NULL)
}
\arguments{
\item{grd}{A \code{SpatialPolygonsDataFrame} object from
\code{grd_build} function.}

\item{pts}{A \code{SpatialPointsDataFrame} object.}

\item{contain}{A \code{character} object \code{simple} or \code{properly}.
By default \code{simple}, it implies a simple contain evaluation (see
\code{\link[rgeos]{gContains}}). \code{properly} implies a properly
contain evaluation (see \code{\link[rgeos]{gContainsProperly}}).}

\item{colnames}{\code{NULL} or \code{character} vector. If \code{NULL}
colnames for the resulting matrix are created as a sequence of the form
F1, F2, \ldots. If \code{character} vector, colnames are
used to name the columns of the resulting matrix.}
}
\value{
An object of class BinMat. See \code{\link[rspecies]{BinMat-class}}
    documentation.
}
\description{
Take a \code{SpatialPointsDataFrame} object and
    compare each points with the boundaries of each cell in a
    grid (\code{SpatialPolygonsDataFrame} from \code{grd_build}
    function.
}
\details{
This function creates a binary matrix with N rows associated to the
    cells ID of \code{grd} object and M columns associated to the factors in
    the \code{pts} object. The element of the row i and column j is 1 if the
    cell i of the \code{grd} object contains at least one point of the factor
    j of the \code{pts} object and 0 otherwise.

    The matrix is created using the functions \code{\link[rgeos]{gContains}}()
    or \code{\link[rgeos]{gContainsProperly}}() from \code{rgeos} package.
    \code{\link[base]{sapply}}() function is used to extract the cell's ids from
    the returned object of \code{\link[rgeos]{gContains}} or \code{gContainsProperly}.
}
\examples{

library(sp)
library(rgeos)

data(Mex0)
data(mammals)

# Generating de grid from Mex0 data
Mex0.grd<-grd_build(Mex0)

# Identification points of mammals
system.time(x.mat<-id_pts(grd = Mex0.grd, pts = mammals))

# Plot sample density - target default
plot(Mex0.grd, x.mat,  sou_den = TRUE, leaflet = TRUE)

# Plot sample density - target "Linx rufus"
plot(Mex0.grd, x.mat, target="Lynx rufus", sou_den = TRUE, leaflet = TRUE)

# Plot sample - target default
plot(Mex0.grd, x.mat,  sou_den = FALSE, leaflet = TRUE)

# Plot sample - target "Linx rufus"
plot(Mex0.grd, x.mat, target="Lynx rufus", sou_den = FALSE, leaflet = TRUE)


}
\author{
Enrique Del Callejo Canal (\email{edelcallejoc@gmail.com}),
    based on implemented algortihms in web platform SPECIES (see References).
}
\references{
\url{http://species.conabio.gob.mx/}
}
\seealso{
\code{\link[rspecies:plot-methods]{plot}}
}

