% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-methods.R
\docType{methods}
\name{plot-methods}
\alias{plot-BinMat}
\alias{plot-BinMatPred}
\alias{plot-methods}
\title{An S4 method for SpatialPolygonsDataFrame and BinMat classes objects.}
\usage{
\S4method{plot}{SpatialPolygonsDataFrame,BinMat}(x, y, target = NULL,
  sou_den = TRUE, leaflet = FALSE, ...)

\S4method{plot}{SpatialPolygonsDataFrame,BinMatPred}(x, y, target = NULL,
  leaflet = FALSE, ...)
}
\arguments{
\item{x}{A \code{SpatialPolygonsDataFrame} object from
\code{\link[rspecies]{grd_build}} function.}

\item{y}{A \code{\linkS4class{BinMat}} object from
\code{\link[rspecies]{id_pts}} function or \code{\linkS4class{BinMatPred}}
object from \code{\link[rspecies]{predict}} function .}

\item{target}{\code{logical}, \code{character}, \code{numeric} or \code{NULL}.
A vector with names, position (integer or boolean) that especifies the
columns of slots DMNB or BMNB from \code{\linkS4class{BinMat}}
that should be considered as target.}

\item{leaflet}{logical. If \code{TRUE} the function uses the leaflet package
and methods. If \code{FALSE} the plotting method from sp package are used.}

\item{...}{Parameters passed to other functions (See
\code{\link[graphics]{par}}).}

\item{sou.den}{logical. If \code{TRUE} the color palette is based on DMNB
slot from \code{\linkS4class{BinMat}} object. If \code{FALSE} the color
palette is based in BMNB slot from \code{\linkS4class{BinMat}} object.}
}
\description{
A plot methods for classes "SpatialPolygonsDataFrame", "BinMat".

A plot methods for classes "SpatialPolygonsDataFrame", "BinMatPred".
}
\examples{
library(sp)
library(rgeos)
data(Mex0)
data(mammals)

# Generating de grid from Mex0 data
Mex0.grd<-grd_build(Mex0)

# Identification points of mammals with colnames = NULL.
x.mat<-id_pts(grd = Mex0.grd, pts = mammals, colnames = NULL)

# plot - BinMat Class
plot(Mex0.grd, x.mat, leaflet = TRUE)

# Counting matrices
count.mat<-counts(x.mat,  target = c("F.169", "F.272"))

# Probability matrices
prob.mat<-probs(count.mat, lap_fac = 0.1)

# score function
score.mat<-score(prob.mat, count.mat)

# Prediction
pred.mat<-predict(score.mat, apr_inc = FALSE, comp_inc = FALSE)

# plot - BinMatPred class
plot(Mex0.grd, pred.mat, target = "Lynx rufus", leaflet = T)

}
\author{
Enrique Del Callejo Canal (\email{edelcallejoc@gmail.com}).
}

