% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\docType{methods}
\name{predict-methods}
\alias{predict-methods}
\alias{predict,BinMatScore}
\title{Function to calculate predictions form BinMatScore Class}
\usage{
\S4method{predict}{BinMatScore}(object, apr_inc = TRUE, comp_inc = TRUE,
  ...)
}
\arguments{
\item{object}{A \code{\link[rspecies]{BinMatScore}} object from
\code{\link[rspecies]{score}} function.}

\item{apr_inc}{logical. If TRUE the apriori score is included for
prediction calculation.}

\item{comp_inc}{logical. If TRUE the complement of explanatory variables
included for prediction claculation}
}
\description{
This function allows to calculate the classifier of each cell in the
    grid of the polygon defined for spatial data mining.
}
\examples{
library(sp)
library(rgeos)
data(Mex0)
data(mammals)

# Generating de grid from Mex0 data
Mex0.grd<-grd_build(Mex0)

# Identification points of mammals with colnames = NULL.
x.mat<-id_pts(grd = Mex0.grd, pts = mammals, colnames = NULL)

# Counting matrices
count.mat<-counts(x.mat,  target = c("F.169", "F.272"))

# Probability matrices
prob.mat<-probs(count.mat, lap_fac = 0.1)

#score function
score.mat<-score(prob.mat, count.mat)

# Prediction
pred.mat<-predict(score.mat, apr_inc = FALSE, comp_inc = FALSE)

# See map with leaflet - "Lynx rufus"

plot(Mex0.grd, pred.mat, target = "Lynx rufus", leaflet = TRUE)

# See data with DT package
library(DT)
datatable(data.frame(pred.mat@Prediction[,head(1:ncol(pred.mat@Prediction))]),
    colnames = pred.mat@name_ID[colnames(pred.mat@Prediction),],
    caption = "laplace factor: 0.1")\%>\%
    formatRound(1:6,digits = 3)

}
\author{
Enrique Del Callejo-Canal (\email{edelcallejoc@gmail.com}).
}
