% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\docType{methods}
\name{score}
\alias{score}
\alias{score,SpNaBaProbs-method}
\alias{score,SpNaba-methods,score-methods}
\title{Score function for the Spatial Naive Bayes Model}
\usage{
\S4method{score}{SpNaBaProbs}(x.probs)
}
\arguments{
\item{x.probs}{A \code{\link[rspeciesdev]{SpNaBaProbs}} object.}
}
\value{
\code{score}: An object of class \code{\link[rspeciesdev]{SpNaBaScore}}.
}
\description{
\code{score} is a function to estimates score parameter for
    the Spatial Naive Bayes Model.
}
\examples{
# Using the function id_pts() for spatial data --------

library(sp)
library(rgeos)
library(rgdal)
library(raster)

# Loading data
data(Mex0)
data(mammals)

# Generating de grid from Mex0 data
Mex0.grd<-grd_build(Mex0)

# Identification points of mammals
x.mat<-id_pts(grd = Mex0.grd, pts = mammals)

# Without target declaration --------------------------

system.time(x.counts <- counts(x.mat))

# With target declaration -----------------------------

x.counts <- counts(x.mat, target = 1:10)

# Probability matrices -----------------------------------
x.probs <- probs(x.counts, fac.lap = 0.1)

# Score parameter -----------------------------------

x.score <- score(x.probs)

}
\references{
\url{http://species.conabio.gob.mx/}
}
\seealso{
\code{\link[rspeciesdev]{SpNaBaProbs}}, \code{\link[rspeciesdev]{SpNaBa-methods}}
}
\author{
Enrique Del Callejo Canal (\email{edelcallejoc@gmail.com}),
    based on implemented algortihms in web platform SPECIES (see References).
}
