% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set-methods.R
\docType{methods}
\name{set}
\alias{set}
\alias{setN}
\alias{setNc}
\alias{setNcx}
\alias{setNx}
\title{Replace methods for BinMat class and subclass}
\usage{
setN(object) <- value

setNc(object) <- value

setNx(object) <- value

setNcx(object) <- value
}
\arguments{
\item{object}{A \code{\link[rspecies]{BinMatCount}} object.}
}
\value{
return a slot of BinMat object.
}
\description{
The setters methods for BinMatCount-class, each function is a
    method to acces to each slot in BinMat-class
}
\author{
Enrique Del Callejo Canal (\email{edelcallejoc@gmail.com}).
}

