% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpNaBaPred.R
\docType{class}
\name{SpNaBaPred-class}
\alias{SpNaBaPred-class}
\alias{initialize,SpNaBaPred-method}
\alias{SpNaBaPred}
\alias{show,SpNaBaPred-method}
\title{Class SpNaBaPred}
\usage{
SpNaBaPred(Prediction, ...)
}
\arguments{
\item{Prediction}{a matrix. See \code{predict}.}

\item{...}{pass to other methods.}
}
\value{
An object of class \code{SpNaBaPred}. See slot section for more details.
}
\description{
This is a class associated to prediction methods applied to
    SpNaBaModel objects (See \code{\link[rspeciesdev]{predict}}).
}
\section{Slots}{

\describe{
\item{\code{Prediction}}{a matrix object.}
}}

\seealso{
\code{\link[rspeciesdev]{grd_build}}, \code{\link[rspeciesdev]{id_pts}}
    \code{\link[rspeciesdev]{NBModel}}, \code{\link[rspeciesdev]{predict}}
}
\author{
Enrique Del Callejo Canal (\email{edelcallejoc@gmail.com}).
}
